app/requirements.txt
flask==3.0.2
tensorflow==2.17.0
numpy==1.26.4
pillow==10.4.0